requirements.txt
appdirs==1.4.0
Django==1.10.5
django-braces==1.11.0
django-cors-middleware==1.3.1
django-dotenv==1.4.1
django-extensions==1.7.7
django-oauth-toolkit==0.11.0
djangorestframework==3.5.3
netifaces==0.10.5
oauthlib==1.1.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
six==1.10.0
